requirements.txt
python-dotenv
aiogram
pandas
openpyxl
